Rename and move some things around.
