+++ .gitignore
@@ -194,4 +194,4 @@ FakesAssemblies/
 
 # Visual Studio 6 workspace options file
 *.opt
-SmugSharpTest/TestAuthentication.cs
+SmugSharpTest/Configuration/Authentication.cs
