--- SmugSharpTest/TestAuthentication.cs
@@ -1,12 +1,0 @@
-﻿namespace SmugSharpTest
-{
-    public static class TestAuthentication
-    {
-        // Fill these in with your own values.
-        public static readonly string ApiKey = "";
-        public static readonly string ApiSecret = "";
-        public static readonly string CallbackUrl = "";
-        public static readonly string AccessToken = "";
-        public static readonly string AccessTokenSecret = "";
-    }
-}
